Add serial driver
Adds driver and matching printer for interacting with Zebra printers
using a serial connection.

The driver is derived from the common ZebraLabelPrintingBaseDriver and
the printer is a port of the standard ZPL library Printer with the
additional SDG functionality of TCPPrinterEnhanced printer.

The main requirement of the serial printer is that frames are wrapped in
start-of-text (<STX> / 0x02) and end-of-text (<ETX> / 0x03) control
characters.

fix serial